+++ rocAL/source/loaders/audio/audio_loader.cpp
@@ -1,5 +1,5 @@
 /*
-Copyright (c) 2019 - 2023 Advanced Micro Devices, Inc. All rights reserved.
+Copyright (c) 2024 Advanced Micro Devices, Inc. All rights reserved.
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -20,15 +20,15 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#include "loaders/image/image_loader.h"
+#include "loaders/audio/audio_loader.h"
+#include "loaders/audio/audio_read_and_decode.h"
 
 #include <chrono>
 #include <thread>
 
-#include "loaders/image/image_read_and_decode.h"
-#include "vx_ext_amd.h"
+#ifdef ROCAL_AUDIO
 
-ImageLoader::ImageLoader(void *dev_resources) : _circ_buff(dev_resources),
+AudioLoader::AudioLoader(void* dev_resources) : _circ_buff(dev_resources),
                                                 _swap_handle_time("Swap_handle_time", DBG_TIMING) {
     _output_tensor = nullptr;
     _mem_type = RocalMemType::HOST;
@@ -36,64 +36,53 @@ ImageLoader::ImageLoader(void *dev_resources) : _circ_buff(dev_resources),
     _output_mem_size = 0;
     _batch_size = 1;
     _is_initialized = false;
-    _remaining_image_count = 0;
+    _remaining_audio_count = 0;
     _device_id = 0;
 }
 
-ImageLoader::~ImageLoader() {
+AudioLoader::~AudioLoader() {
     de_init();
 }
 
-void ImageLoader::shut_down() {
+void AudioLoader::shut_down() {
     if (_internal_thread_running)
         stop_internal_thread();
     _circ_buff.release();
 }
 
-void ImageLoader::set_prefetch_queue_depth(size_t prefetch_queue_depth) {
+void AudioLoader::set_prefetch_queue_depth(size_t prefetch_queue_depth) {
     if (prefetch_queue_depth <= 0)
-        THROW("Prefetch quque depth value cannot be zero or negative");
+        THROW("Prefetch queue depth value cannot be zero or negative");
     _prefetch_queue_depth = prefetch_queue_depth;
 }
 
-void ImageLoader::set_gpu_device_id(int device_id) {
+void AudioLoader::set_gpu_device_id(int device_id) {
     if (device_id < 0)
         THROW("invalid device_id passed to loader");
     _device_id = device_id;
 }
 
 size_t
-ImageLoader::remaining_count() {
-    if (_external_source_reader) {
-        if ((_image_loader->count() != 0) && !_external_input_eos)
-            return _batch_size;
-        else {
-            return 0;
-        }
-    }
-    return _remaining_image_count;
+AudioLoader::remaining_count() {
+    return _remaining_audio_count;
 }
 
-void ImageLoader::reset() {
+void AudioLoader::reset() {
     // stop the writer thread and empty the internal circular buffer
     _internal_thread_running = false;
     _circ_buff.unblock_writer();
-
     if (_load_thread.joinable())
         _load_thread.join();
-
     // Emptying the internal circular buffer
     _circ_buff.reset();
-
     // resetting the reader thread to the start of the media
-    _image_counter = 0;
-    _image_loader->reset();
-
+    _audio_counter = 0;
+    _audio_loader->Reset();
     // Start loading (writer thread) again
     start_loading();
 }
 
-void ImageLoader::de_init() {
+void AudioLoader::de_init() {
     // Set running to 0 and wait for the internal thread to join
     stop_internal_thread();
     _output_mem_size = 0;
@@ -102,21 +91,16 @@ void ImageLoader::de_init() {
 }
 
 LoaderModuleStatus
-ImageLoader::load_next() {
-    return update_output_image();
+AudioLoader::load_next() {
+    return update_output_audio();
 }
 
-void ImageLoader::set_output(Tensor *output_tensor) {
+void AudioLoader::set_output(Tensor* output_tensor) {
     _output_tensor = output_tensor;
-    _output_mem_size = ((_output_tensor->info().data_size() + 8) & ~7);  // Making output size as a multiple of 8 to support vectorized load and store in RPP
+    _output_mem_size = _output_tensor->info().data_size();
 }
 
-void ImageLoader::set_random_bbox_data_reader(std::shared_ptr<RandomBBoxCrop_MetaDataReader> randombboxcrop_meta_data_reader) {
-    _randombboxcrop_meta_data_reader = randombboxcrop_meta_data_reader;
-    _circ_buff.random_bbox_crop_flag = true;
-}
-
-void ImageLoader::stop_internal_thread() {
+void AudioLoader::stop_internal_thread() {
     _internal_thread_running = false;
     _stopped = true;
     _circ_buff.unblock_reader();
@@ -126,97 +110,82 @@ void ImageLoader::stop_internal_thread() {
         _load_thread.join();
 }
 
-void ImageLoader::initialize(ReaderConfig reader_cfg, DecoderConfig decoder_cfg, RocalMemType mem_type, unsigned batch_size, bool decoder_keep_original) {
+void AudioLoader::initialize(ReaderConfig reader_cfg, DecoderConfig decoder_cfg, RocalMemType mem_type, unsigned batch_size, bool decoder_keep_original) {
     if (_is_initialized)
         WRN("initialize() function is already called and loader module is initialized")
-
     if (_output_mem_size == 0)
-        THROW("output image size is 0, set_output() should be called before initialize for loader modules")
-
+        THROW("output audio size is 0, set_output() should be called before initialize for loader modules")
     _mem_type = mem_type;
     _batch_size = batch_size;
     _loop = reader_cfg.loop();
-    _decoder_keep_original = decoder_keep_original;
-    _image_loader = std::make_shared<ImageReadAndDecode>();
+    _audio_loader = std::make_shared<AudioReadAndDecode>();
     size_t shard_count = reader_cfg.get_shard_count();
     int device_id = reader_cfg.get_shard_id();
     try {
         // set the device_id for decoder same as shard_id for number of shards > 1
-        if (shard_count > 1)
-            _image_loader->create(reader_cfg, decoder_cfg, _batch_size, device_id);
-        else
-            _image_loader->create(reader_cfg, decoder_cfg, _batch_size);
-    } catch (const std::exception &e) {
+        if (shard_count > 1) {
+            _audio_loader->Create(reader_cfg, decoder_cfg, _batch_size, device_id);
+        } else {
+            _audio_loader->Create(reader_cfg, decoder_cfg, _batch_size);
+        }
+    } catch (const std::exception& e) {
         de_init();
         throw;
     }
-    _max_tensor_width = _output_tensor->info().max_shape().at(0);
-    _max_tensor_height = _output_tensor->info().max_shape().at(1);
-    _decoded_data_info._data_names.resize(_batch_size);
-    _decoded_data_info._roi_height.resize(_batch_size);
-    _decoded_data_info._roi_width.resize(_batch_size);
-    _decoded_data_info._original_height.resize(_batch_size);
-    _decoded_data_info._original_width.resize(_batch_size);
-    _crop_image_info._crop_image_coords.resize(_batch_size);
+    _max_decoded_samples = _output_tensor->info().max_shape().at(0);
+    _max_decoded_channels = _output_tensor->info().max_shape().at(1);
+    _decoded_audio_info._data_names.resize(_batch_size);
+    _decoded_audio_info._original_audio_samples.resize(_batch_size);
+    _decoded_audio_info._original_audio_channels.resize(_batch_size);
+    _decoded_audio_info._original_audio_sample_rates.resize(_batch_size);
     _circ_buff.init(_mem_type, _output_mem_size, _prefetch_queue_depth);
     _is_initialized = true;
-    _image_loader->set_random_bbox_data_reader(_randombboxcrop_meta_data_reader);
     LOG("Loader module initialized");
 }
 
-void ImageLoader::start_loading() {
+void AudioLoader::start_loading() {
     if (!_is_initialized)
         THROW("start_loading() should be called after initialize() function is called")
 
-    _remaining_image_count = _image_loader->count();
+    _remaining_audio_count = _audio_loader->Count();
     _internal_thread_running = true;
-    _load_thread = std::thread(&ImageLoader::load_routine, this);
+    _load_thread = std::thread(&AudioLoader::load_routine, this);
 }
 
 LoaderModuleStatus
-ImageLoader::load_routine() {
+AudioLoader::load_routine() {
     LOG("Started the internal loader thread");
     LoaderModuleStatus last_load_status = LoaderModuleStatus::OK;
-    // Initially record number of all the images that are going to be loaded, this is used to know how many still there
+    // Initially record number of all the audios that are going to be loaded, this is used to know how many still there
 
     while (_internal_thread_running) {
-        auto data = _circ_buff.get_write_buffer();
+        auto data = reinterpret_cast<float*>(_circ_buff.get_write_buffer());
         if (!_internal_thread_running)
             break;
 
         auto load_status = LoaderModuleStatus::NO_MORE_DATA_TO_READ;
         {
-            load_status = _image_loader->load(data,
-                                              _decoded_data_info._data_names,
-                                              _max_tensor_width,
-                                              _max_tensor_height,
-                                              _decoded_data_info._roi_width,
-                                              _decoded_data_info._roi_height,
-                                              _decoded_data_info._original_width,
-                                              _decoded_data_info._original_height,
-                                              _output_tensor->info().color_format(), _decoder_keep_original);
+            load_status = _audio_loader->Load(data,
+                                              _decoded_audio_info,
+                                              _max_decoded_samples,
+                                              _max_decoded_channels);
 
             if (load_status == LoaderModuleStatus::OK) {
-                if (_randombboxcrop_meta_data_reader) {
-                    _crop_image_info._crop_image_coords = _image_loader->get_batch_random_bbox_crop_coords();
-                    _circ_buff.set_crop_image_info(_crop_image_info);
-                }
-                _circ_buff.set_decoded_data_info(_decoded_data_info);
+                _circ_buff.set_decoded_data_info(_decoded_audio_info);
                 _circ_buff.push();
-                _image_counter += _output_tensor->info().batch_size();
+                _audio_counter += _output_tensor->info().batch_size();
             }
         }
         if (load_status != LoaderModuleStatus::OK) {
             if (last_load_status != load_status) {
                 if (load_status == LoaderModuleStatus::NO_MORE_DATA_TO_READ ||
                     load_status == LoaderModuleStatus::NO_FILES_TO_READ) {
-                    LOG("Cycled through all images, count " + TOSTR(_image_counter));
+                    LOG("Cycled through all audios, count " + TOSTR(_audio_counter));
                 } else {
-                    ERR("ERROR: Detected error in reading the images");
+                    ERR("ERROR: Detected error in reading the audios");
                 }
                 last_load_status = load_status;
             }
-
             // Here it sets the out-of-data flag and signal the circular buffer's internal
             // read semaphore using release() call
             // , and calls the release() allows the reader thread to wake up and handle
@@ -230,19 +199,19 @@ ImageLoader::load_routine() {
     return LoaderModuleStatus::OK;
 }
 
-bool ImageLoader::is_out_of_data() {
-    return (remaining_count() < _batch_size);
+bool AudioLoader::is_out_of_data() {
+    return (remaining_count() < 0);
 }
+
 LoaderModuleStatus
-ImageLoader::update_output_image() {
+AudioLoader::update_output_audio() {
     LoaderModuleStatus status = LoaderModuleStatus::OK;
 
     if (is_out_of_data())
         return LoaderModuleStatus::NO_MORE_DATA_TO_READ;
     if (_stopped)
         return LoaderModuleStatus::OK;
-
-    // _circ_buff.get_read_buffer_x() is blocking and puts the caller on sleep until new images are written to the _circ_buff
+    // _circ_buff.get_read_buffer_x() is blocking and puts the caller on sleep until new audios are written to the _circ_buff
     if ((_mem_type == RocalMemType::OCL) || (_mem_type == RocalMemType::HIP)) {
         auto data_buffer = _circ_buff.get_read_buffer_dev();
         _swap_handle_time.start();
@@ -258,27 +227,23 @@ ImageLoader::update_output_image() {
     }
     if (_stopped)
         return LoaderModuleStatus::OK;
-
-    _output_decoded_data_info = _circ_buff.get_decoded_data_info();
-    if (_randombboxcrop_meta_data_reader) {
-        _output_cropped_img_info = _circ_buff.get_cropped_image_info();
-    }
-    _output_names = _output_decoded_data_info._data_names;
-    _output_tensor->update_tensor_roi(_output_decoded_data_info._roi_width, _output_decoded_data_info._roi_height);
+    _output_decoded_audio_info = _circ_buff.get_decoded_data_info();
+    _output_names = _output_decoded_audio_info._data_names;
+    _output_tensor->update_tensor_roi(_output_decoded_audio_info._original_audio_samples, _output_decoded_audio_info._original_audio_channels);
+    _output_tensor->update_audio_tensor_sample_rate(_output_decoded_audio_info._original_audio_sample_rates);
     _circ_buff.pop();
     if (!_loop)
-        _remaining_image_count -= _batch_size;
-
+        _remaining_audio_count -= _batch_size;
     return status;
 }
 
-Timing ImageLoader::timing() {
-    auto t = _image_loader->timing();
+Timing AudioLoader::timing() {
+    auto t = _audio_loader->GetTiming();
     t.process_time = _swap_handle_time.get_timing();
     return t;
 }
 
-LoaderModuleStatus ImageLoader::set_cpu_affinity(cpu_set_t cpu_mask) {
+LoaderModuleStatus AudioLoader::set_cpu_affinity(cpu_set_t cpu_mask) {
     if (!_internal_thread_running)
         THROW("set_cpu_affinity() should be called after start_loading function is called")
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32) && !defined(__CYGWIN__)
@@ -291,7 +256,7 @@ LoaderModuleStatus ImageLoader::set_cpu_affinity(cpu_set_t cpu_mask) {
     return LoaderModuleStatus::OK;
 }
 
-LoaderModuleStatus ImageLoader::set_cpu_sched_policy(struct sched_param sched_policy) {
+LoaderModuleStatus AudioLoader::set_cpu_sched_policy(struct sched_param sched_policy) {
     if (!_internal_thread_running)
         THROW("set_cpu_sched_policy() should be called after start_loading function is called")
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32) && !defined(__CYGWIN__)
@@ -303,20 +268,11 @@ LoaderModuleStatus ImageLoader::set_cpu_sched_policy(struct sched_param sched_po
     return LoaderModuleStatus::OK;
 }
 
-std::vector<std::string> ImageLoader::get_id() {
+std::vector<std::string> AudioLoader::get_id() {
     return _output_names;
 }
 
-DecodedDataInfo ImageLoader::get_decode_data_info() {
-    return _output_decoded_data_info;
+DecodedDataInfo AudioLoader::get_decode_data_info() {
+    return _output_decoded_audio_info;
 }
-
-CropImageInfo ImageLoader::get_crop_image_info() {
-    return _output_cropped_img_info;
-}
-
-void ImageLoader::feed_external_input(const std::vector<std::string>& input_images_names, const std::vector<unsigned char *>& input_buffer, const std::vector<ROIxywh>& roi_xywh, unsigned int max_width, unsigned int max_height, unsigned int channels, ExternalSourceFileMode mode, bool eos) {
-    _external_source_reader = true;
-    _external_input_eos = eos;
-    _image_loader->feed_external_input(input_images_names, input_buffer, roi_xywh, max_width, max_height, channels, mode, eos);
-}
+#endif